Slice parametrisation - stage 2
What I've done here is create a parametrised fixture function called
input_expout_combinator that takes parametrised inputs of slice objects
which are then used to slice the input and expected output dataframes.
These dataframes are also passed to input_expout_combinator as
additional arguments since they're used in the function body. The sliced
dataframes are then returned as a tuple.

For the test, the new fixture function is passed as an argument and the
input and expected output dataframes are unpacked from the tuple. Since
input_expout_combinator is a parametrised fixture function, the test
runs for each set of parameters. So in this case the test will run four
times each with a different slice of both the input and expout dataframe
(given by the parameters).

Additional notes:

* The fourth parameter is wrapped in a pytest.param() function to enable
  the addition of a mark. This was so historically I was able to add an
  xfail mark for a slice of the dataframe that the function didn't solve
  yet. The function given is it's final version though, and will pass
  this case - but thought I'd leave this in as a demonstration.
* I've chosen to mark each slice of the test data with a comment to
  indicate which slice is for each case.
* I've added a form of BDD notation in comments under each id string to
  describe the details of each test.
* There are some assertion statements in the fixture function just to
  check that the end of the slices given actually exist in the
  dataframe.