Kwargs parametrisation - stage 1
This method of parametrisation allows us to parametrise existing
fixtures. Furthermore, it allows us to include only the parameters we
need for that case, in others words we can include a variable amount of
parameters to the test. This is useful if we'd rather use the default
arguments for a parameter on one test, and pass a value on another test.

The key to achieving this is making use of the `kwargs` attribute of the
`Case` class which is a dictionary of all the named parameters passed to
a `Case` instance (excluding `label` which is a separate attribute in
itself). Therefore, the `input_expout_combinator` fixture function
returns a dict mapping of {name: value} for our parameters. In the test
body, the expout parameter is popped from the dictionary to its own
variable. This then allows the unpacking of the remaining
`input_expout_combinator` keyword arguments into our function under
test. This use of the Python operator `**` for unpacking keyword
arguments is what enables us to pass a variable number of parameters to
our function under test.

Additional notes:

* In it's current format, the fixture function only accepts other
  fixtures as parameters, since it is using the
  `request.getfixturevalue()` function to return the value of that
  fixture parameter inside the fixture function body. This would break
  if the fixture didn't exist. Notice that the fixtures are passed in
  as strings. The following commit will show how it can be generalised
  to take a mix of functions and standard data types.