Kwargs parametrisation 2 and mixed parametrisation
This commit includes an update to the kwargs_parametrisation.py module
and another complete example, mixed_parametrisation.py because they are
related. They both include the abstracted function `get_case_parameters`
which is located in helprs.py (you may want to put it in conftest.py
depending on preference).

The `get_case_parameters` is a flexible function that requires a
FixtureRequest object to be passed to it. It also expects that the
FixtureRequest contains an instance of the `Case` class for each
parameter i.e. passed to `params` in the fixture decorator.

So what `get_case_parameters` does, is it will return a dictionary of
parameters given by the instance of `Case`, and if the parameter is a
fixture (in string format) it will return the value of that fixture, and
if the the parameter is not a fixture, then it will just return the
value passed. This allows you to pass any number of fixtures and
non-fixtures to a test for a specific case. Neat huh? The
`get_case_parameters` function makes use of a `get_fixture_value`
function which just uses a try except clause to attempt to get the
fixture value.

You'll notice the format for both files committed here is exactly the
same - so this provides a consistent structure of test for many
different use cases. Have fun with it!