Multiple parameter parametrisation - stage 2
This was my first attempt at parametrising the different test cases for
`filter_retailer_items`. I took the approach of creating a fixture for
each of the different expected outputs, and then combining the parameter
for the filter and the expected output in the `param_expout_combinator`
fixture function.

I was a little unhappy with this approach though, because the parameters
passed to the `params` in the decorator of the fixture function were not
named so it's not immediately clear what they are. Additionally, the
fixtures for each of the expected output dataframes are quite simple -
there's not much setup, just a call of `create_dataframe` - so having
each in a fixture is probably over-engineered.

The next commit show how this was tidied up a bit.