Multiple parameter parametrisation - stage 3
The final stage here resulted in the use of the helper function
`parametrize_cases, which is simply a wrapper around the existing
pytest.mark.parametrize decorator within pytest but provides a more
reader friendly way of submitting parameters. The motivation behind
using this ahead of the what's already available in pytest is explained
further by the author here:

https://github.com/ckp95/pytest-parametrize-cases

What this has enabled us to do in this use case, is to define each
parameter by name, co-located with a "label" for each test case which is
much more readable. The benefit of using parametrize_cases as opposed to
creating a fixture function and using `get_case_parameters` which I've
used in some of the other examples, is that you don't need to create the
additional fixture function - you can ust decorate the test and pass in
the parameters by name.

A big disadvantage though is that it doesn't work with fixtures -
there's no way to obtain the fixture value. It works for this example
because the expected output dataframes were simple enough to list in a
readable way within each call of the `Case` class - and we were able to
get rid of all the additional function boiler plate for each fixture -
which was a gain for readability.

It also wouldn't work for a variable number of parameters unless None is
explicitly passed to each parameter not being used in a specific test
case.